Essais Bit

+++ Core/include/DBinary.h
@@ -42,9 +42,9 @@ using namespace std;
 
 #ifdef USE_64BIT_IDS
 //     typedef UINT8 INT_TYPE;
-    typedef UINT64 BIN_TYPE;
+    typedef UINT8 BIN_TYPE;
 #else    
-    typedef UINT32 BIN_TYPE;
+    typedef UINT8 BIN_TYPE;
 #endif // USE_64BIT_IDS 
 
 struct bitIndex
